:sparkles: add calculator and replace it
