+++ lib/main/calendar/calendar.dart
@@ -1,3 +1,4 @@
+import 'package:Pilll/main/calendar/calculator.dart';
 import 'package:Pilll/main/record/weekday_badge.dart';
 import 'package:Pilll/model/weekday.dart';
 import 'package:Pilll/theme/font.dart';
@@ -9,19 +10,9 @@ abstract class CalendarConstants {
 }
 
 class Calendar extends StatelessWidget {
-  final DateTime date;
-  const Calendar({Key key, this.date}) : super(key: key);
+  final Calculator calculator;
 
-  DateTime _dateTimeForFirstDayOfMonth() {
-    return DateTime(date.year, date.month, 1);
-  }
-
-  int _lastDay() => DateTime(date.year, date.month + 1, 0).day;
-  int _weekdayOffset() =>
-      WeekdayFunctions.weekdayFromDate(_dateTimeForFirstDayOfMonth()).index;
-  int _previousMonthDayCount() => _weekdayOffset();
-  int _tileCount() => _previousMonthDayCount() + _lastDay();
-  int _lineCount() => (_tileCount() / 7).ceil();
+  const Calendar({Key key, this.calculator}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -38,25 +29,27 @@ class Calendar extends StatelessWidget {
                   )),
         ),
         Divider(height: 1),
-        ...List.generate(_lineCount(), (line) {
+        ...List.generate(calculator.lineCount(), (line) {
           return Column(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
               Row(
                 children: Weekday.values.map((weekday) {
                   bool isPreviousMonth =
-                      weekday.index < _weekdayOffset() && line == 0;
+                      weekday.index < calculator.weekdayOffset() && line == 0;
                   if (isPreviousMonth) {
                     return CalendarDayTile(
                         disable: true,
                         weekday: weekday,
-                        day: _dateTimeForPreviousMonthTile(weekday).day);
+                        day: calculator
+                            .dateTimeForPreviousMonthTile(weekday)
+                            .day);
                   }
                   int day = line * CalendarConstants.weekdayCount +
                       weekday.index -
-                      _weekdayOffset() +
+                      calculator.weekdayOffset() +
                       1;
-                  bool isNextMonth = day > _lastDay();
+                  bool isNextMonth = day > calculator.lastDay();
                   if (isNextMonth) {
                     return Expanded(child: Container());
                   }
@@ -72,17 +65,6 @@ class Calendar extends StatelessWidget {
         }),
       ],
     );
-  }
-
-  DateTime _dateTimeForPreviousMonthTile(Weekday weekday) {
-    var dateTimeForLastDayOfPreviousMonth = DateTime(date.year, date.month, 0);
-    var offset =
-        WeekdayFunctions.weekdayFromDate(dateTimeForLastDayOfPreviousMonth)
-            .index;
-    return DateTime(
-        dateTimeForLastDayOfPreviousMonth.year,
-        dateTimeForLastDayOfPreviousMonth.month,
-        dateTimeForLastDayOfPreviousMonth.day - offset + weekday.index);
   }
 }
